+++ yash-syntax/CHANGELOG.md
@@ -41,6 +41,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Internal dependency versions
     - thiserror 1.0.47 → 2.0.4
 
+### Changed
+
+- Error messages returned from `parser::SyntaxError::message` are no longer
+  capitalized.
+
 ## [0.12.1] - 2024-11-10
 
 ### Changed